Use ASSERT0 to check locations in derivatives
Most of the derivatives throw BoutException if the location of
argument(s) and outloc are inconsistent. A few were using ASSERT1 to
check locations, so didn't throw if checking was 'off'. It's more
consistent to throw always, so use ASSERT0 instead. (Allows
tests/MMS/difops to treat all derivatives the same.)